authorization

+++ nestjs-server/src/tasks/tasks.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, ParseIntPipe, Patch, Post, Req, Res, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, ParseIntPipe, Patch, Post, Req, Res, UnauthorizedException, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
 import { Request, Response } from 'express';
 import { TasksService } from './tasks.service';
 import { title } from 'process';
@@ -14,33 +14,59 @@ export class TasksController {
     
    @Get()
    @UseGuards(AuthGuard)
-   async getTasks(@Param() params:any , @Res() res:Response):Promise<void> {
+   async getTasks(@Param() params:any , @Res() res:Response , @Req() req : Request):Promise<void> {
+         if(!req?.user?.userId)
+         {
+           throw new UnauthorizedException("Not authorized");
+         }    
         const tasks= await this.tasksService.getAllTasks();
         res.status(200).json({tasks});
    }
 
    @Post("create")
    @UsePipes(ValidationPipe)
-   async createNewTask(@Body() createTasksDto : createTaskDTO , @Res() res: Response) : Promise<void> {
-    const newTask = await this.tasksService.createTask(createTasksDto);
+   @UseGuards(AuthGuard)
+   async createNewTask(@Body() createTasksDto : any , @Req() req: Request , @Res() res: Response) : Promise<void> {
+       const creator = req.user.userId;
+       if(!req?.user?.userId)
+       {
+         throw new UnauthorizedException("Not authorized");
+       }   
+    const newTask = await this.tasksService.createTask({creator , ...createTasksDto});
      res.status(201).json({newTask});
    }
-
+   
    @Get(":id")
-  async  getTaskById(@Param("id" ) id:string , @Res() res:Response) :Promise<void> {
+   @UseGuards(AuthGuard)
+  async  getTaskById(@Param("id" ) id:string , @Res() res:Response , @Req() req:Request) :Promise<void> {
+    if(!req?.user?.userId)
+    {
+      throw new UnauthorizedException("Not authorized");
+    }   
     const task = await this.tasksService.getTaskById(id);
      res.status(200).json({task});
    }
 
    @Delete(":id")
-   async deleteTask(@Param("id" ) id:string , @Res() res:Response) :Promise<void> {
+   @UseGuards(AuthGuard)
+   async deleteTask(@Param("id" ) id:string , @Res() res:Response ,  @Req() req:Request) :Promise<void> {
+    if(!req?.user?.userId)
+    {
+      throw new UnauthorizedException("Not authorized");
+    }   
     const task = await this.tasksService.deleteTask(id);
      res.status(201).json({message : task});
    }
 
    @Patch(":id/status")
-   async updateTaskStatus(@Param("id") id:string , @Body("status" , TaskStatusValidation) status : TaskStatus , @Res() res:Response): Promise<void> {
-       const updatedTask = await this.tasksService.updateTaskStatus(id ,  status);
+   @UseGuards(AuthGuard)
+   async updateTaskStatus(@Param("id") id:string , @Body("status" , TaskStatusValidation) status : TaskStatus 
+   , @Res() res:Response ,  @Req() req:Request): Promise<void> {
+    if(!req?.user?.userId)
+    {
+      throw new UnauthorizedException("Not authorized");
+    }   
+    const updatedTask = await this.tasksService.updateTaskStatus(id ,  status);
         res.status(201).json({updatedTask});
    }
 }
